Create Thor vs Loki
The universe contains a magic number z. Thor's power is known to x and Loki's power to be y.
One's strength is defined to be z - a, if his power is a. Your task is to find out who among Thor and Loki has the highest strength, and print that strength.